Adds ability to return results as JSON
Returning results as JSON makes testing easier by allowing state
(tweets) to be stored in a serialized format and then mocked in without touching
Twitter.

There was an issue with the JSONParser in tweepy so I switched to the
twitter module.  Right now data.py will use the twitter module if
JSON is requested and tweepy if a stream is requested.  I think that
it makes sense to move compeletely to the twitter module, but I don't
want to make that decision for you.

Now that search results can be stored in files such as
/tests/fixtures/tweet.json, repeatable tests can be created to test the
search functionality.